이미지 삭제 추가

+++ src/service/upload.service.js
@@ -1,4 +1,5 @@
 const { executeQuery } = require("../repository");
+const fs = require('fs');
 
 const uploadService = {
   uploadFile: async (req, res) => { // 등록
@@ -12,21 +13,21 @@ const uploadService = {
         url: path
       };
 
-      const sql = "INSERT INTO tbl_file SET ?";
+      const sql = `INSERT INTO tbl_file SET ?`;
       const saveAnswer = await executeQuery(sql, data);
 
       res.status(200).json({
         code: 200,
-        message: "컨텐츠 등록에 성공하였습니다.",
+        message: '이미지 등록에 성공하였습니다.',
         data: { file_id: saveAnswer.insertId },
       });
     }catch (err) {
       let message;
 
       if(req.session.user == undefined) {
-        message = "유저 세션이 존재하지 않습니다."
+        message = '유저 세션이 존재하지 않습니다.'
       } else {
-        message = "에러가 발생하였습니다.";
+        message = '에러가 발생하였습니다.';
       }
 
       console.error(err);
@@ -35,33 +36,49 @@ const uploadService = {
         .json({ message: message, data: err, code: 500 });
     }
   },
-  deleteFile: async (req, res) => { // 등록
+  deleteFile: async (req, res) => { // 삭제
     try {
-      const { originalname, filename, path, mimetype } = req.file;
+      const { file_id } = req.body;
+
+      let sql = `UPDATE tbl_file SET del_yn = 'Y' WHERE file_id = ?`;
+      let response = await executeQuery(sql, [file_id]);
+      
+      sql = `SELECT * FROM tbl_file WHERE file_id = ?`;
+      response = await executeQuery(sql, [file_id]);
 
       // 업로드된 파일의 경로
-      const filePath = req.file.path;
+      let filePath;
+      if(response[0]) {
+        filePath = response[0].url;
 
-      // 파일 삭제
-      fs.unlink(filePath, (err) => {
-        let message;
-
-        if(err) {
-          message = "파일 삭제 실패"
-        } else {
-          message = "파일 삭제 성공";
-        }
-
-        console.error(err);
-        res
-          .status(500)
-          .json({ message: message, data: err, code: 500 });
-      });
-    } catch {
+        // 파일 삭제
+        fs.unlink(filePath, (err) => {
+          if (err) {
+            console.error(err);
+            res.status(500).json({
+              code: 500,
+              message: '이미지 삭제 실패',
+              data: err
+            });
+          } else {
+            res.status(200).json({
+              code: 200,
+              message: '이미지 삭제 성공',
+              data: {filePath: filePath}
+            });
+          }
+        });
+      } else {
+        res.status(200).json({
+          code: 500,
+          message: '해당 이미지 미존재'
+        });
+      }
+    } catch (err) {
       console.error(err);
       res
         .status(500)
-        .json({ message: '파일 삭제 실패', data: err, code: 500 });
+        .json({ message: '이미지 삭제 실패', data: err, code: 500 });
     }
   },
 };
